requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
mysqlclient==2.2.4
gunicorn==22.0.0
